requirements-prod.txt
# Production Web Server
gunicorn==21.2.0

# Flask Framework
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# Environment & Config
python-dotenv==1.0.0

# Payment Processing
stripe==8.0.0

# AI Services
google-generativeai==0.3.2
replicate==0.22.0
openai==1.12.0

# Image Processing
Pillow==10.2.0
# rembg==2.0.56  # Optional, removed for production deployment (requires C dependencies)

# HTTP & Utilities
requests==2.31.0

# Production Optimization
brotli==1.1.0
